move some test file to match src code
